qa-lab: share evidence gallery view types
